+++ web-client/src/modules/requests/pages/routes/FindRequestsRoute/FindRequestsRoute.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import FindRequestsContainer from '../../../../home/containers/CAV_HOME/FoundRequestsContainer';
+import FindRequestsContainer from '../../../containers/FoundRequests_CAV_HOME/FoundRequestsContainer';
 
 const FindRequestsRoute: React.FC = () => <FindRequestsContainer />;
